crash.types.percpu: Add percpu handling
This commit adds support for special handling of percpu variables.  It will
accept symbols that are in the percpu sections, symbols that point to percpu
objects, and passed pointers to percpu objects.  If a symbol or value is
passed that is not a percpu (as determined by whether the offset falls
within the percpu section), a TypeError is raised.

Thanks to Vlastimil Babka for the initial work in extending percpu handling
to values instead of just symbols.

Test cases are also included to test a variety of access methods.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>