crash: add support for active threads
The kernel data structures in the crash dump tell us which threads were
active when the dump was taken but not the contents of those threads'
registers.  The dump format does contain that information so we need
to access it from the vmcore attributes instead.

This commit also fixes some API issues that remained from a previous
version of libkdumpfile.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>