crash.commands: add help command
`pyhelp' will now list all of the commands implemented via crash.commands

It will also display the help text for an individual command if specified.

Since the help text is integrated into the gdb.Command stuff as well,
'help <command>' also works, as does 'help user-defined'

Using  after the first help text line makes it so the summary appears
in 'help user-defined' but we get the full text without the truncated
header when getting help for a particular command.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>