arch/x86_64: do not fetch rip unless specifically requested
We try to batch register updates as a performance optimization, but
updating 'rip' has a side effect: it will reinitialize the frame cache.

When this happens during 'frame info', it causes a segfault.

So, now we just update rip when it's requested specifically or if
all registers are requested.  Since it's not setting up the thread
anymore, we'll rename the affected routines to fetch_register_*.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>