types/util: support nested struct and unions in offsetof
There was a problem with list_for_entry_each() on struct slab, which looks
like this:

struct slab {
        union {
                struct {
                        struct list_head list;
                        unsigned long colouroff;
                        void *s_mem;            /* including colour offset */
                        unsigned int inuse;     /* num of objs active in slab */
                        kmem_bufctl_t free;
                        unsigned short nodeid;
                };
                struct slab_rcu __slab_cover_slab_rcu;
        };
};

The problem boils down to offsetof() trying to do a slab_type["list"] raising
KeyError, where slab_type is gdb.Type for struct slab. Interestingly enough,
slab["list"] on a gdb.Value object works fine.

Presumably the Type object should work as well as the Value object, but for
now implement a workaround with recursive offsetof() for nested structs and
unions.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>