types/list: add cycle detection
Detect cycles in linked lists and abort traversal, using Floyd's Tortoise and
Hare algorithm. Because we have double-linked lists, cycles that don't include
the initial head are only possible if there's a broken prev pointer somewhere,
so it's sufficient to start the cycle detection only when we see a broken prev
pointer.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>