crash.types.util: make offsetof much more robust
This commit rewrites offsetof to handle a number of cases it would either
fail with or produce unexpected results.

The fix to handle anon structs and unions also introduced a bug where
it would hunt for members even in named structs so that things like:
inode.wait_list would find inode.i_mutex.wait_list, which isn't what we
want.

This also adds support for things like explicitly requesting the offset
of embedded members like inode.i_mutex.wait_list so that we don't need
to reinvent that for the struct command.

This also adds a unit test infrastructure that exercises offsetof with
45 new tests, but also allows us to add tests for other parts of the
code.  The test cases for offsetof use a small standalone C program
to provide the object file.  Hopefully, later cases can do the same.

To run the tests, just execute ./test-all.sh from the repo root.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>