crash.commands.struct: rework and fix struct
Now that we have the 'struct test' object file for testing
crash.types.util, we can use it to test crash.commands.struct as well.

This commit adds a bunch of test cases and reworks the command to
be more easily tested.  It also contains fixes for the test failures.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>