crash.infra: add @delayed_init, @exporter, and @export decorators
These three decorators codify how code should be structured and exported
when establishing crash-python infrastructure modules.

Classes that require information from the underlying executable or target
must be declared with the @delayed_init decorator.  This delays execution
of the __init__ method until the first attribute that needs initialization
is accessed.  This is critical for implementing unit tests so that
importing modules doesn't attempt to pull in the world as it starts up.

Modules that publish an API, which should be all of them, must locate
all routines in a class declared with the @exporter decorator.  Each
method to be exported to the module-level must be decorated with
@export.  The @exporter decorator instantiates the class as a singleton
and "steals" the class name as the instance name.  It then locates
any methods declared with @export and publishes them in the module
namespace.  Instance, class, and static methods are all supported.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>