crash.infra: enforce decorator order
If @exporter is applied before @delayed_init, it will end up triggering
the delayed initialization during import, which is exactly what these
are intended to avoid.

It should be possible to make the ordering not matter but for now let's
just throw a RuntimeError to avoid weird bugs like sizeof(void *) == 4
when builtin types are looked up before we load the executable.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>