crash.infra: fix delayed_init and exporter, add real testing
Further development showed that the crash.infra implementation was
incomplete and hacky.  This patch converts delayed_init to a class that
the decorator creates using the original class as a base class.  It
also converts the exporter to use the inspect interface, which does
do what we need and uses bound methods so we don't need to play games
with the wrapper.

Lastly, if @exporter is applied before @delayed_init, it will end up
triggering the delayed initialization during import, which is exactly
what these are intended to avoid.  We'll throw a TypeError in this case
instead of just causing weird bugs.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>