crash: add support for loading modules and debuginfo
This commit adds support for loading kernel modules into the address
space.  It will search a list of directories recursively looking for
modules and load them automatically.  This is a change over legacy
crash but is necessary since we have full stack unwinding and we need
the unwind information from the modules to make that work.

In order to use any of this, we need valid debuginfo.  For self-built
kernels, this will typically still be in the object files but for
kernels we ship, the debuginfo is separate.  We'll search the same
directories for the debuginfo and load it automatically as well.

We also add a simple driver to load everything up.  The easiest
way to use crash is now the crash.sh script.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>