crash.infra: document limitation with super().__init__() and delayed_init
There is a major shortcoming with delayed_init inheriting from the base
class behind the scenes: super() will refer to the base class.  As a
result, super().__init__() call from base_class.__init__() will
result in infinite recursion.

It's fixable, but it's also easy enough to work around.  We'll leave it
as a problem to solve on another day.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>