crash.commands: make more testable
We currently do a few things that makes it difficult to do testing of
commands:
- The command parser exits, which gdb catches and turns into a SystemExit
  exception, but this could mean anything.
- We catch a few exceptions ourselves, so even if we called
  CrashCommand.invoke directly instead of gdb.execute(), we won't actually
  get an normal error condition back -- only the exceptional ones.

This patch moves the command execution into a new invoke_uncaught routine
that doesn't catch exceptions.  The old invoke continues to catch
exceptions, so we don't fall over during normal execution.

We have a new CrashCommandParser that /must/ be used when creating new
commands.  This is so we can raise our own exception that we can
handle in the unit tests.  All existing commands are switched over

We register the command as the last step so we don't trip over the
TypeError later.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>