crash.types.task: use type from value rather than lookup
GDB's type lookup doesn't use the symbol cache so we get different but
identical struct types backing the lookup and the type associated
with a value.  Since the pointers differ, a deep comparison needs to
be made to see if they are identical.  Since task_struct references
a huge number of other structures via pointers, the deep comparison
is very expensive. When loading thousands of tasks, this can be very
time consuming.

We can work around the issue by swapping the lookup type for the value
type when we have a matching one.

We also raise a new exception for bad tasks since we're raising and
formatting it in two places.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>