crash: use build directory for in-repo execution
The crash.sh and test-all.sh scripts use the repo directory for the
module path.  This can result in modules that existed at one point
but were discarded to be used if the .pyc files still exist.

This patch removes the build/lib/crash directory, rebuilds it with
setup.py build, and uses that for runtime and testing.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>