crash.infra: add symbol/type callback infrastructure
GDB provides the ability to execute callbacks when objfiles are
loaded.  Rather than blindly delay initialization, we can subscribe
to the callback events and initialize the types and symbols as they
are loaded.

Using the SymbolCallback or TypeCallback classes, the consumer can
have a callback fire once when the symbol or type is available.  A
specific mechanism to assign the result to an attribute in the caller
is also provided.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>