crash.subsystem.storage: Add storage subsystem infrastructure
This patch adds a set of modules for use with the storage subsystem.

Initially, it provides a bio chain unwinder that allows storage
implementations and file systems to register callbacks that decode
a specific bio.  Basic device mapper support is provided.

This also pulls in crash.types.blockdev and eliminates the old module.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>