crash.infra: exporter, create instance on first use
Replacing the class name with an instance was troublesome and
unnecessary.  Now we keep a dictionary of the singleton instances
attached to the module and use a callable wrapper to instantiate
the singleton when required.  This should also translate into needing
to use delayed initialization much less since the instances aren't
created during import - only the wrappers are.  It also means that
the ordering with @exporter and @delayed_init doesn't matter anymore.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>