crash.infra: replace @exporter with new base class
The @exporter decorator is a hacky way to export to the module namespace.

This commit adds a new metaclass that does the same thing and adds
a base class that uses that metaclass for everything else to inherit from.

Now all that's required is to inherit from the base class and decorate
using @export as before.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>