crash.kdump.target: Re-introduce symbol callback
The current code works just fine if correct information can be found in
the dump file itself (using VMCOREINFO). However, if this information is
not available, address translation may fail.

The new code still prefers data from VMCOREINFO, because that is more
likely to match the dump file, but if the default method fails, it
consoluts GDB symbol table.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>