vtop: Reuse libkdumpfile address translation if possible
The address translation system is currently created from scratch
using built-in heuristics and data from gdb. However, there is
already a translation system when working with a libkdumpfile
target.

BTW this is the only way to make address translation work on some
dump files, e.g. Xen xc_core ELF dump files contain the mapping
between physical and machine addresses in the non-data portion of
the file. In other words, it is not mapped to any virtual address,
so the only way to reach it is through direct access to the dump
file.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>