crash.command.dmesg: Fix to work again
The "charp" type is 32-bit from some reason. A cast to this type
might cause losing the higher 32 bits of the address.

It helps to use the delayed lookup for the type.

This patch also removes two debug messages that printed
"msg" and "textval". I guess that someone used them to
debug this problem.

Signed-off-by: Petr Mladek <pmladek@suse.com>