crash.session: Load kernel symbols with kASLR offset
With kASLR, all symbols are shifted by an offset (which can be
found in VMCOREINFO). GDB already has some internal support for
such shifting (to cope with user-space ASLR), but it cannot be
re-used here, because it depends on infrastructure that does not
exist in the kernel, or even makes no sense.

Instead, I'm using the ability to relocate sections at load time.
There are a few peculiarities with this approach:

  1. This feature is available only with the "add-symbol-file"
     command, not with the "symbol-file" command.
  2. Each relocated section must be explicitly listed.
  3. The section list is not available through any Python
     interface.

The solution is to set the exec-file first (to get all section
names and deal with issue 2 above), then parse the output of "info
target" (to deal with issue 3 above) and finally invoke the
"add-symbol-file" command with appropriate arguments.

Ugly, as it is, this approach works without further changes in gdb
itself.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>