crash.sh: support for separate gdb installation
At this stage of development, unpatched gdb doesn't support the features
that crash-python requires.  Rather than expect the user to replace
their gdb with a patched on, we can install a separate gdb and use that.

We'll check for crash-python-gdb first and then fall back to gdb.  We also
will now throw an error if no gdb is found.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>