crash: add support for kASLR
This patch enables kASLR support for crash-python.  libkdumpfile does
the heavy lifting for address translation but the devil is in the details.

There are a few ordering issues that needed to be sorted out:
- The base address comes from the vmcore
- If we load the kernel without the base address, all of the objfile
  callbacks for delayed lookups will be called with the unrelocated
  addresses
- add-symbol-file doesn't set the architecture, so we need to do that
  ourselves.

Petr Tesarik's work on gdb to support an objfile-wide offset makes this
work quite a bit easier.  One hiccup is that, internally, gdb takes
the section offset into account when generating addresses for minimal
symbols.  This means that providing an address for .text that already
has the offset applied doesn't work because the minimal symbols will
not be properly relocated.  If it's omitted, the -o functionality does
the right thing.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>