crash.subsystem.storage: use unqualified types when passed by value
When a type is passed via a value or symbol, it can be const or
volatile.  While the latter is rare in the kernel, the former is
commonplace and is one of the things that many subsystems add to
operations structures and the like over time.

This patch ignores the qualifications when doing a type comparison.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>