types/percpu: add KASLR support
The relocation done by KASLR adjusts values of symbols, including percpu ones
which are not real pointers, but offsets. This is compensated by subtracting
the __per_cpu_start which is also adjusted. However percpu values that are not
symbols (both static and dynamic) remain unchanged, so the subtraction of
__per_cpu_start yields wrong result. We have to detect the KASLR offset
(currently x86_64-specific hack) and subtract it for non-symbols.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>