UNTESTED: crash.subsystem.storage.device-mapper: update for device mapper end_io changes
Linux commit bfc6d41cee53b (dm: stop using bi_private) landed in v3.15,
which means that resolving the context from bio->bi_private would fail.

This commit updates the symbol callbacks to test for the 'clone' member
of each of the contexts and assigns a getter method to get the context
in each decoder.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>