crash.sh, crash, contrib: Fix except syntax for python3
In python3, the python2 syntax "except exc, var" has been changed into
"except exc as var".  For python 2.6, the new "as" syntax is accepted as an
alternative to the "," syntax.

This patch updates the remaining "," occurances to the "as" syntax, fixing the
syntax for python3.

While this means that those occurances no longer work with pre-python2.6,
there are already 25 occurances of the "as" syntax in crash-python, so it's
safe to assume that pre-python2.6 is already not supported.