crash.infra.lookup: Fix formatting for python3
In python3, indentation is rejected as inconsistent if a source file mixes
tabs and spaces in a way that makes the relative indentation level of
subsequent lines dependent on the tab size.

This fixes a "TabError: inconsistent use of tabs and spaces in indentation"
in crash/infra/lookup.py by using indentation on the offending line that is
consistent with the surrounding lines.