crash.util: add struct_has_member
The use of anonymous structures and unions means that things like:

struct foo {
	struct {
		int x;
	};
};

if 'x' in cls.foo_type:
	# will evaluate false

when foo.x works fine in C code.  In order to make these less painful
for subsystem modules, we add a struct_has_member helper that does
the right thing to resolve the member.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>