crash.types.percpu: get_percpu_var, raise exception with passed val
If the value passed to get_percpu_var isn't a percpu, we raise an
exception -- but the exception contained the processed value instead
of the passed one.  That can be misleading when debugging.

Also, handle the val=None case that can occur as we try to treat the
value as a pointer.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>