crash.util.symbols: introduce simplified delayed lookups
The CrashBaseClass method of resolving types and symbols automatically
is overcomplicated.  We have classes in modules that have no use
for them other than to access the type and symbol resolution.  We then
have to export helpers to the module namespace for convenience.  It's
not obvious to most people why any of this stuff is the way it is, which
makes contributing a challenge.

This commit introduces a new crash.utils.symbol module that offers a
simplified way to do it.  We still have the automatic importing into
a namespace, but the namespaces are not embedded in a class.  Instead,
we have a DelayedCollection object that is extended for particular uses.

For example, we can convert for_each_super_block to the following:

from crash.util.symbols import Types, Symvals

types = Types(['struct super_block', 'struct module *'])
symvals = Symvals(['super_blocks'])

def for_each_super_block():
    for sb in for_each_list_entry(symvals.super_blocks, types.super_block_type,
                                  s_list)
        yield sb

Unfortunately, moving the attribute creation outside of the metaclass
means that it's subject to Python name mangling.  This means that
symbols prefixed with two or more underscores cannot be accessed via
the attribute "dot" interface.  The DelayedCollection class supports
attribute access via the dict "[name]" interface as well as via a
get(name) accessor.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>