crash: remove CrashBaseClass
This commit removes CrashBaseClass from the project entirely.

This simplifies the code immensely and means contributors don't need
to understand and/or potentially debug the black magic behind the
namespace, implicit singleton, exporting, and type/symbol loading
in order to get started.

The pattern now is to use crash.types.symbols to declare a Types,
Symbols, Symvals, etc object at the top of the module that contains
any needed types or symbols.  These are still declared as an array
of names and the naming structure (e.g. struct_name_type) is the
same as with CrashBaseClass.  They just appear in separate collections
under separate names.  Most functions will be at the module level
and will reference these objects directly.  If a state object is
required, a class is still used but it can inherit from `object' instead.
If callbacks are used on class attributes (instead of instance attributes)
they must be declared outside of the class.

This allows us to eliminate many classes entirely and reduce others
to simple state holder classes.  In some cases, like with percpu,
it makes sense to just use module instance and marshall the calls from the
module level to the instance explicitly.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>