crash: add kernel testing against vmcores
One of the things that has had a serious negative effect on the quality
of crash-python is the inability to do automated testing on a broad
variety of kernels and vmcores.  Often, we have to run our tests
by hand.  More often, we get reports from new users that something
broke unexpectedly.

This commit adds real unit testing against real kernels and vmcores.
It includes a few test cases that will need to be extended further.

The policy moving forward will be that new features will require
a matching test case that passes across a variety of kernels.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>