crash.commands: explicity define help text using ::format_help
This commit stops (ab)using the class docstring for the command
to contain help text.  There was nothing inherently wrong with doing it
but it means that the entire command help text appears in the API
documentation and needs to be properly formatted rst, which isn't what
we want or need.

Commands now provide a ::format_help method that returns a string
containing the help text.  All leading and trailing whitespace will
be stripped prior to printing so we have some flexibility as far as
how to declare the strings.  The class docstring will now only
contain the summary, which is more suitable for inclusion in the API
documentation.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>