tests: clean up testing
This commit spits up test-all.sh into individual scripts accessible via
make rules.  We also add exit rules for the unit tests so that failures
are actually passed back out to the calling script.

At this point only the unit tests and live tests are run as part of
'make tests.'  Once the static checks and lint errors/warnings are
cleared, we'll enable those as well.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>