vtop: don't instantiate CrashAddressTranslation on module import
The CrashAddressTranslation constructor requires a valid syscache
be set up prior to running.  This isn't a problem generally but
when it's constructed during module import, it means unit tests
that just import the module fail.  Since it's only used for
the vtop command and initialization is lightweight, move it
to the vtop execute method.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>