docs: create gdb.inv prior to apidoc
intersphinx and our apidoc handler both run as builder-inited callbacks,
which means that gdb.inv isn't created for the first run through.

The proper way to create the file is via a config-inited event callback.
This works fine for readthedocs.  My local system uses an earlier version
without that callback so we'll use a make rule to create it instead.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>