lint: disable 'bad-option-value' check
The 'bad-option-value' check checks the "pylint: disable=" comments to
ensure the names are valid.  The problem is that it means older lint
don't work since the valid names aren't known to them yet.

I'd rather throw lint failures on commit hook (when that's implemented)
than force anyone developing crash-python to have the latest tools that
are just for sanity checks.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>