tests: add stubs for gdb types
In order to get full test coverage, we need to provide stubs for
external modules.  We don't use enough of elftools to bother with
the stubs.  libaddrxlat needs surgery in order for it to pass type
checking.  So for now, we just use the typing for gdb.

This means that the static checker will now check the interface
with GDB.  Prior to this change, the interface was not checked at all.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>