crash: fix some typing errors
mypy really doesn't like it when you reuse the same variable for different
things.  The "worst" case is crash.requirements, which is easy to fix.

The other cases are just errors that didn't get caught due to gdb not
having type stubs.  I'm writing those up and they caught a few errors.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>