crash.types.task: remove TASK_EXCLUSIVE
This bit was originally cloned from crash and, as best I can tell, it
hasn't been used in any kernel in nearly 20 years.  There was also
a bug associated with declaring all the flags where this one is handled
specially and using -1 as a placeholder meant it always matched.

As a result, state would always = 0 and all tasks would be shown as running.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>