kdump.target: accept both vmlinux and vmcore arguments
GDB 9.1 is leaving exec-file symbols behind even when symbol-file
symbols are loaded at an offset.  As a result, we get the original
unrelocated address and encounter memory errors.

This change passes the vmlinux path to the kdumpfile target so
that we can get the base offset prior to loading the kernel.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>