crash.commands.kmem: keep working even if system uses CONFIG_SLUB
In Yocto 5.8 kernel it was observed that pycrash produces the following
exception:

...
  File "/home/skamensk/crash-python/20201101/crash-python/build/lib/crash/commands/kmem.py", line 28, in <module>
    from crash.types.slab import kmem_cache_get_all, kmem_cache_from_name
...
  File "/home/skamensk/crash-python/20201101/crash-python/build/lib/crash/types/slab.py", line 328, in check_kmem_cache_type
    cls.percpu_name = find_member_variant(gdbtype, ['cpu_cache', 'array'])
  File "/home/skamensk/crash-python/20201101/crash-python/build/lib/crash/util/__init__.py", line 304, in find_member_variant
    raise TypeError("Unrecognized '{}': could not find member '{}'"
TypeError: Unrecognized 'struct kmem_cache': could not find member 'cpu_cache'

Yocto Project 5.8 Kernel configured with CONFIG_SLUB. In this case struct
kem_cache does not have field cpu_cache, so lookup of this field fails in
such configuration.

Workaround is to check kernel config if CONFIG_SLUB is enabled and skip slab
specif functionality if that is the case.

TODO: Implement slub analogs for slab operations.

Signed-off-by: Alexander Kamensky <alexander.kamensky42@gmail.com>