crash.arch.x86_64: make stack traces more reliable
We can't depend on CFI in __switch_to_asm, but we can unwind just
enough to where it was called and use the saved register state to
set up the frame.  This caches the stack frame size (to adjust rsp)
and the location of the caller of __switch_to_asm to set up the frames
for other scheduled tasks.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>