crash-python: Update to use gdb-12.1 with kdumpfile target
The crash-python GDB implementation has been updated to use gdb 12.1,
which integrates libkdumpfile into libbfd and offers a small semantic
target to improve performance.

As part of the update, the registers interface has changed, we populate
the thread list from the target implementation in C, and use the core
file interface.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>